Extracted kiz functions to extra module.
